+++ test-data/src/main/resources/org/openmrs/module/fhir2/providers/AllergyIntolerance_json_merge_patch.json
@@ -1,11 +1,5 @@
 {
-  "clinicalStatus": {
-    "coding": [
-      {
-        "system": "http://terminology.hl7.org/CodeSystem/allergyintolerance-clinical",
-        "code": "active"
-      }
-    ],
-    "text": "Active"
-  }
+  "category": [
+    "food"
+  ]
 }
