+++ test-data/src/main/resources/org/openmrs/module/fhir2/providers/AllergyIntolerance_json_patch.json
@@ -1,0 +1,7 @@
+[
+  {
+    "op": "replace",
+    "path": "/category/0",
+    "value": "food"
+  }
+]
